+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"/Users/Dilan/Desktop/projects/summerlinlabs/src/components/Admin/AdminSchedule.js":"1","/Users/Dilan/Desktop/projects/summerlinlabs/src/App.js":"2","/Users/Dilan/Desktop/projects/summerlinlabs/src/apiKey.js":"3","/Users/Dilan/Desktop/projects/summerlinlabs/src/components/Login.js":"4"},{"size":1434,"mtime":1606540309169,"results":"5","hashOfConfig":"6"},{"size":1101,"mtime":1606451304084,"results":"7","hashOfConfig":"6"},{"size":360,"mtime":1606445760799,"results":"8","hashOfConfig":"6"},{"size":1933,"mtime":1606540457431,"results":"9","hashOfConfig":"6"},{"filePath":"10","messages":"11","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"vev3m7",{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"16","messages":"17","errorCount":0,"warningCount":2,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},"/Users/Dilan/Desktop/projects/summerlinlabs/src/components/Admin/AdminSchedule.js",[],"/Users/Dilan/Desktop/projects/summerlinlabs/src/App.js",[],"/Users/Dilan/Desktop/projects/summerlinlabs/src/apiKey.js",[],"/Users/Dilan/Desktop/projects/summerlinlabs/src/components/Login.js",["18","19"],{"ruleId":"20","severity":1,"message":"21","line":10,"column":11,"nodeType":"22","messageId":"23","endLine":10,"endColumn":16},{"ruleId":"20","severity":1,"message":"24","line":24,"column":10,"nodeType":"22","messageId":"23","endLine":24,"endColumn":16},"no-unused-vars","'login' is assigned a value but never used.","Identifier","unusedVar","'userID' is assigned a value but never used."]
+[{"/Users/Dilan/Desktop/projects/summerlinlabs/src/components/Admin/AdminSchedule.js":"1","/Users/Dilan/Desktop/projects/summerlinlabs/src/App.js":"2","/Users/Dilan/Desktop/projects/summerlinlabs/src/apiKey.js":"3","/Users/Dilan/Desktop/projects/summerlinlabs/src/components/Login.js":"4","/Users/Dilan/Desktop/projects/summerlinlabs/src/components/NavBar.js":"5","/Users/Dilan/Desktop/projects/summerlinlabs/src/pages/LoginPage.js":"6","/Users/Dilan/Desktop/projects/summerlinlabs/src/pages/Contact.js":"7","/Users/Dilan/Desktop/projects/summerlinlabs/src/components/BackgroundImage/TestCatalog.js":"8","/Users/Dilan/Desktop/projects/summerlinlabs/src/components/ContactForm.js":"9","/Users/Dilan/Desktop/projects/summerlinlabs/src/contexts/AuthContext.js":"10","/Users/Dilan/Desktop/projects/summerlinlabs/src/index.js":"11","/Users/Dilan/Desktop/projects/summerlinlabs/src/pages/Mission.js":"12","/Users/Dilan/Desktop/projects/summerlinlabs/src/components/Wrapper/index.js":"13","/Users/Dilan/Desktop/projects/summerlinlabs/src/pages/Homepage.js":"14","/Users/Dilan/Desktop/projects/summerlinlabs/src/components/Footer.js":"15","/Users/Dilan/Desktop/projects/summerlinlabs/src/components/Dashboard/Dashboard.js":"16"},{"size":1659,"mtime":1607044175158,"results":"17","hashOfConfig":"18"},{"size":973,"mtime":1607043983135,"results":"19","hashOfConfig":"18"},{"size":360,"mtime":1606445760799,"results":"20","hashOfConfig":"18"},{"size":1907,"mtime":1607039638145,"results":"21","hashOfConfig":"18"},{"size":2370,"mtime":1607043971380,"results":"22","hashOfConfig":"18"},{"size":1422,"mtime":1607044339464,"results":"23","hashOfConfig":"18"},{"size":247,"mtime":1606237690300,"results":"24","hashOfConfig":"18"},{"size":8492,"mtime":1606237690294,"results":"25","hashOfConfig":"18"},{"size":2000,"mtime":1606423960090,"results":"26","hashOfConfig":"18"},{"size":1369,"mtime":1606451746056,"results":"27","hashOfConfig":"18"},{"size":358,"mtime":1606237690299,"results":"28","hashOfConfig":"18"},{"size":278,"mtime":1606237690300,"results":"29","hashOfConfig":"18"},{"size":163,"mtime":1606237690298,"results":"30","hashOfConfig":"18"},{"size":322,"mtime":1606451592900,"results":"31","hashOfConfig":"18"},{"size":1470,"mtime":1606237690295,"results":"32","hashOfConfig":"18"},{"size":1237,"mtime":1607045206585,"results":"33","hashOfConfig":"18"},{"filePath":"34","messages":"35","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"vev3m7",{"filePath":"36","messages":"37","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"38","messages":"39","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"40","messages":"41","errorCount":0,"warningCount":2,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},{"filePath":"42","messages":"43","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"44","messages":"45","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"46","messages":"47","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"48","messages":"49","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"50","messages":"51","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"52","messages":"53","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"54","messages":"55","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"56","messages":"57","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"58","messages":"59","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"60","messages":"61","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"62","messages":"63","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"64","messages":"65","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/Users/Dilan/Desktop/projects/summerlinlabs/src/components/Admin/AdminSchedule.js",[],"/Users/Dilan/Desktop/projects/summerlinlabs/src/App.js",[],"/Users/Dilan/Desktop/projects/summerlinlabs/src/apiKey.js",[],"/Users/Dilan/Desktop/projects/summerlinlabs/src/components/Login.js",["66","67"],"/Users/Dilan/Desktop/projects/summerlinlabs/src/components/NavBar.js",[],"/Users/Dilan/Desktop/projects/summerlinlabs/src/pages/LoginPage.js",[],"/Users/Dilan/Desktop/projects/summerlinlabs/src/pages/Contact.js",[],"/Users/Dilan/Desktop/projects/summerlinlabs/src/components/BackgroundImage/TestCatalog.js",[],"/Users/Dilan/Desktop/projects/summerlinlabs/src/components/ContactForm.js",[],"/Users/Dilan/Desktop/projects/summerlinlabs/src/contexts/AuthContext.js",[],"/Users/Dilan/Desktop/projects/summerlinlabs/src/index.js",[],"/Users/Dilan/Desktop/projects/summerlinlabs/src/pages/Mission.js",[],"/Users/Dilan/Desktop/projects/summerlinlabs/src/components/Wrapper/index.js",[],"/Users/Dilan/Desktop/projects/summerlinlabs/src/pages/Homepage.js",[],"/Users/Dilan/Desktop/projects/summerlinlabs/src/components/Footer.js",[],"/Users/Dilan/Desktop/projects/summerlinlabs/src/components/Dashboard/Dashboard.js",[],{"ruleId":"68","severity":1,"message":"69","line":5,"column":9,"nodeType":"70","messageId":"71","endLine":5,"endColumn":13},{"ruleId":"68","severity":1,"message":"72","line":24,"column":10,"nodeType":"70","messageId":"71","endLine":24,"endColumn":16},"no-unused-vars","'auth' is defined but never used.","Identifier","unusedVar","'userID' is assigned a value but never used."]
