+++ src/components/Dashboard/style.css
@@ -1,0 +1,3 @@
+.admin-container .see-all-schedule {
+    background-color: #2bbbad !important;
+}
